refactor(gemini): 将 GeminiModalContent 内联到 GeminiSettings 并移除独立文件
